https://tenor.com/view/smart-iam-the-smartest-man-alive-gif-8196999 like literally
